# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for BPM.",
        "title": "BPM API",
        "contact": {
            "name": "Lewis",
            "email": "lxx0103@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
        "/apis": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API管理"
                ],
                "summary": "API列表",
                "operationId": "36",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "API路由",
                        "name": "route",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/auth.API"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API管理"
                ],
                "summary": "新建API",
                "operationId": "37",
                "parameters": [
                    {
                        "description": "API信息",
                        "name": "api_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.APINew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.API"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/apis/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API管理"
                ],
                "summary": "根据ID获取API",
                "operationId": "38",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.API"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API管理"
                ],
                "summary": "根据ID更新API",
                "operationId": "39",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "API信息",
                        "name": "api_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.APINew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.API"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
        "/menuapis/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据菜单ID获取API权限",
                "operationId": "46",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据菜单ID更新API权限",
                "operationId": "47",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单信息",
                        "name": "menu_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.MenuAPINew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.API": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "auth.APINew": {
            "type": "object",
            "required": [
                "method",
                "name",
                "route",
                "status"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "post",
                        "put",
                        "get"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "route": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
...
        "response.ErrorRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ListRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "page_id": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
    }
}